moved map of San Francisco to main Dashboard page
